+++ __tests__/__fixtures__/bailout/early-return/output.js
@@ -1,19 +1,23 @@
-const _collection = [1, 2, 3];
-const _fn = (_value) => {
-  if (_value === 2) {
-    return true;
-  }
-};
-const _results = [];
-for (
-  let _key = 0, _length = _collection.length, _value, _result;
-  _key < _length;
-  ++_key
-) {
-  _value = _collection[_key];
-  _result = _fn(_value, _key, _collection);
-  if (_result) {
-    _results.push(_value);
-  }
+function getFoo(config) {
+  const collection = config.collection || [1, 2, 3];
+  return (() => {
+    const _fn = (_value) => {
+      if (_value === 2) {
+        return true;
+      }
+    };
+    const _results = [];
+    for (
+      let _key = 0, _length = collection.length, _value, _result;
+      _key < _length;
+      ++_key
+    ) {
+      _value = collection[_key];
+      _result = _fn(_value, _key, collection);
+      if (_result) {
+        _results.push(_value);
+      }
+    }
+    return _results;
+  })();
 }
-const result = _results;