+++ __tests__/__fixtures__/complex/conditional-if/code.js
@@ -1,8 +1,8 @@
 import { map } from '../../../../src/inline-loops.macro';
 
-function getStuff() {
+function getStuff(array, foo) {
   if (foo === 'bar') {
-    return map(array, v => v * 2);
+    return map(array, (v) => v * 2);
   }
 
   return array;