+++ __tests__/__fixtures__/complex/conditional-ternary/code.js
@@ -1,0 +1,5 @@
+import { map } from '../../../../src/inline-loops.macro';
+
+function getStuff(array, foo) {
+  return foo === 'bar' ? map(array, (v) => v * 2) : array;
+}